Add support for series upgrade
We can't actually pause the service because the series upgrade hooks run
on the subordinates before the principal so doing so could interfere
with the pod drain process. So instead, we just set a status to let the
Juju admin know that it's ok to proceed.

Part of https://bugs.launchpad.net/charm-canal/+bug/186994